+++ test/Veggerby.Boards.Tests/Cards/CardsModuleEdgeTests.cs
@@ -1,12 +1,14 @@
 using System;
-using System.Linq;
 using System.Collections.Generic;
+using System.Linq;
+
+using AwesomeAssertions;
 
 using Veggerby.Boards.Cards;
-using Veggerby.Boards.States;
 using Veggerby.Boards.Flows.Events;
+using Veggerby.Boards.States;
+
 using Xunit;
-using AwesomeAssertions;
 
 namespace Veggerby.Boards.Tests.Cards;